feat: create_room done
