+++ src/messages/constants.js
@@ -6,5 +6,6 @@ export const messageType = {
   ATTACK: 'attack',
   FINISH: 'finish',
   UPDATE_ROOM: 'update_room',
-  UPDATE_WINNERS: 'update_winners'
+  UPDATE_WINNERS: 'update_winners',
+  CREATE_ROOM: 'create_room'
 };
